+++ core/database/migrations/2019_09_28_150239_create_rincians_table.php
@@ -15,7 +15,7 @@ class CreateRinciansTable extends Migration
     {
         Schema::create('rincians', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('permohonan_id')->nullable();
+            $table->bigInteger('permohonan_id')->nullable();
             $table->string('namapermohonan');
             $table->string('jenisbelanja');
             $table->bigInteger('biayasatuan');
